docshell/test/mochitest/test_bug694612.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=694612
-->
<head>
  <title>Test for Bug 694612</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=694612">Mozilla Bug 694612</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 694612 **/
SimpleTest.waitForExplicitFinish();

window.addEventListener("message", receiveMessage);
function receiveMessage(event) {
  ok(event.data.result, "should have performance API in an <object>");
  window.removeEventListener("message", receiveMessage);
  SimpleTest.finish();
}
</script>
<object type="text/html"
        data="data:text/html,<script>parent.postMessage({result:performance!=null},'*');</script>">
</object>
</pre>
</body>
</html>
